Delete new_chapters document as it is rarely used.
Joel says this was a template used to generate a shell when adding a set of
related APIs.  It was written in 1997 and a good portion is untouched.

If we need this functionality in the new docs we can write something much nicer
using Python to generate the proper reST documentation.