Remove ada_user document.
We are going to make Ada a chapter in the c_user manual to simplify things.
Talked it over with Joel the differences between ada_user and c_users are a lot
smaller than they used to be.

A simple few page chapter will be enough for anyone to know the differences and
will be a lot easier to maintain.