waf: Have configure set building pdf and/or singlehtml.
Move selecting pdf and singlehtml to the configure stage so it
is sticky for all builds. This means a top level build will always
build all formats that have been configured.

Do not complete the configure stage if tools are missing for the
configured output.

Add singlehtml support using the inliner tool. It is nice.

Remove the groups as waf can track the dependences. This lets the
manuals build in parallel.