waf: Add support to handle missing Latex packages on hosts they are not available on.
It appears the support for texlive packages on some hosts is variable. This patch
lets us add missing packages to our source tree so a PDF can be built on
those hosts. The quality of the PDFs created may vary as some short cuts may
have been take. For example lato is a font and only the sty file as been added
and not the actual font which means it's use will default to another font.