Initial addition of SPARC V8 Stack/Register Windowing from Peter Magnusson
This content was originally at https://www.sics.se/~psm/sparcstack.html
and is currently only available via the Wayback Machine. Peter kindly gave
permission to include it with the RTEMS Documentation. This initial commit
is a cut and paste with minimal edits. Edits are needed to ensure proper
formatting and the figures need to be made to look correct in the Rest
formatted output.