user: add a warning not to use special characters in pathnames
Spaces and other specical characters in pathnames break builds of any
software packages using GNU make.

All of the GNU tools and RTEMS use make and this breaks builds. This
is a long standing known issue with GNU make, which can't handle spaces
in pathnames:

https://savannah.gnu.org/bugs/?712

This work was part of GCI 2018.

Probably closes #3450.