Fix error messages when Sphinx version cannot be checked.
This is kind of a dual-check since it checks for the version while simultaneously
checking to see if Sphinx exists on the system.

Changed the error messages to be clear on the error and removed a reference to
version which cannot be set due to the check failure.

Reported by: Shashvat Jain