user: Improve bug report section
The new content is based on the wiki page:

https://devel.rtems.org/wiki/Developer/Bug_Reporting

The goal is to remove the wiki page and direct the "New ticket" page to
this user manual section.  Some content of the wiki page should move to
the "Contributing" sections since it is not directly related to bug
reporting, e.g. managing existing bugs.