user/hosts/posix.rst: Update Ubuntu instructions.
Note that the ticket has instructions for what to do if the
error message "Error :: You must put some 'source' URIs in your sources.list"
is reported. As this occurred on an Ubuntu version that was not
LTS, it is not being added. If this error occurs in the future on a
LTS release, then add this information.

closes #3416.