c-user: Remove copyright from Petr Benes
The content introduced by a commit in the RTEMS main repository

    commit 418de420a05609ba8919822b553706963a8d3a7b
    Author: Joel Sherrill <joel.sherrill@OARcorp.com>
    Date:   Wed Oct 5 19:59:47 2011 +0000

    2011-10-05      Joel Sherrill <joel.sherrill@oarcorp.com>
                    Petr Benes <benesp16@fel.cvut.cz>

            PR 1912/doc
            * user/conf.t, user/schedule.t: Rework to add scheduler specific
            information.

is no longer in this file.  According to the file history, this was the
only content introduced by Petr Benes.  Also during the relicensing of
the documentation to CC-BY-SA-4.0 in 2016 it was proclaimed that OAR was
the only copyright holder of the RTEMS documentation present in the
RTEMS main repository.

Update #3836.