posix-compliance: Use constant name for API spreadsheet.
README-cvs.txt will need to be updated each time the spreadsheet is updated.
The spreadsheet is managed and maintained externally and has a version in
the name. We do not want that in the repository because a diff of the CSV
should always be much smaller than a "delete old/add new" patch.