bsps/imxrt: Remove superfluous options from dtc
The -@ option adds symbols that are used when applying an overlay. RTEMS
doesn't support overlays or re-parsing of device trees. So the symbols
only increase the size of the binary.

The same is true for the padding. It would be useful if the device tree
is changed during run time which is not supported in RTEMS.

Update #4180