eng: Add register block specification types
A register block may be used to specify the interface of devices which
contain registers associated with an integer address.  Register blocks
consist of register block members specified by the ``definition``
attribute.  Register block members are either instances of registers
specified by the ``registers`` attribute or instances of other register
blocks specified by links with the "register-block-include" link role.
Registers consists of bit fields.  The register block members are placed
into the address space of the device relative to the base address of the
register block.  Register member offests and the register block size are
specified in units of the address space granule.

Update #3715.