user/bsps/arm: improve stm32h7 BSP description
This patch adds kind of HOWTO to the stm32h7 BSP description. The howto
discusses various tools from STM tool-chain and how to use them in order
to configure board for RTEMS and to upload RTEMS application binary
to the board. The patch also adds more supported boards to the table.

Sponsored-By:	Precidata