bsps/riscv: Remove inaccurate statement about reliance on a boot loader
The BSP is capable of initialising the hardware being the first software
that takes control on hardware reset (after the bootrom). For instance,
using on QEMU's  virt platforms, RTEMS runs as a bios without BBL.
Similarily, RTEMS can also be run on harware/FPGA and loaded using
GDB; the bootrom (or a GDB script) should just set the a0/a1 registers
with the boot HART ID and DTB address respectively.