docs/user: add docs for riscv/kendrytek210 BSP variant
This patch adds the documentation for building and running RTEMS on the Kendryte K210
RISC-V SoC. The generic riscv introducion was re-arranged to list the multilib variants
then the specific hardware targets. In addition a couple of errors were fixed for the
generic QEMU commands.

V2 corrected a typo, expanded K210 Console UART parameters, and addded a hyperlink
to renode.io install instructions.

Closes #4876