Integrate legacy-networking into user manual
RTEMS has multiple addon packages. Adding a separate manual for each
one will lead to a big amount of manuals. That would make the
documentation overview page crowded.

Beneath that, it's much simpler to add a chapter to a manual compared to
adding a new manual. So collecting add-on package manuals as part of the
user manual hopefully will make it simpler to write documentation for
new packages and therefore lead to better documented software.

This commit integrates the existing legacy-networking manual into the
user manual as a first step.