Change theme to sphinx-book-theme
Sphinx booktheme is a more modern theme based off of of the PyData theme for
science.

It is cleaner and more suitable for the RTEMS documentation.  Works very well
for mobile and is easier to navigate for large, complex documents.

This also includes the refreshed logo and favicon

Some other cleanups to conf.py to make it easier to manage.