sphinx: Remove digest strings and fix sidebar
Inliner is old there are no alternatives so remove the ?digest= and ?v= strings
so itcan build the source from the new theme.

Also fix the html_sidebars from the commandline use singlehtml_sidebars.index.
The behaviour of html_sidebars was changed in:

  https://github.com/sphinx-doc/sphinx/pull/12600

It will no longer take a string from the commandline.

Fixes: #57, #58