posix-compliance: Fix building of subpages
MyST does not like more than a single H1 per file it makes sense and is far
better for organising docs.  This rewrites everything to split it into different
files and has to have a deferred build in order to include the generated files
on the sphinx-build commandline

With tons of help from @ita2048 thank you!