c-user/message/directives.md: Note non-atomicity
rtems_message_queue_broadcast() may, under certain circumstances, copy
the message to tasks which were not waiting on the message queue when
the broadcast started, and may copy the message multiple times to the
same task.

This behaviour is, based on discussion in
https://gitlab.rtems.org/rtems/rtos/rtems/-/issues/4804, something that
might not change in the short term, so expose it in the user
documentation.

The wording is copied with modifications from an existing internal
requirements note.

The changes were auto-generated from rtems-central.